Function providing loading for data templates (with single and multiple data sheets) (data should have been cleaned before).

Function providing loading for data templates (with single and multiple data sheets) (data should have been cleaned before).

Usage

load_template(template_path, epsg_code = 4326)

Arguments

template_path

path to *.xlsx file.

epsg_code

EPSG code for station coordinates (default 4326 for WGS 84 coordinate system) system

Value

list of two: stations dataframe, which contains station information, second member of list contains contains measurement data in dataframe (if data is in one sheet) or nested list of dataframes with each parameter in separate dataframe.

Examples

temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
load_template(temp_path)
#> [1] "Loading data from template."
#> [1] "Loading of data is finished."
#> $stations
#> Simple feature collection with 20 features and 5 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 18.58608 ymin: 52.56443 xmax: 18.73169 ymax: 52.65544
#> Geodetic CRS:  WGS 84
#> # A tibble: 20 × 6
#>    ID    Name                      Descr…¹            geometry  Long   Lat
#>  * <chr> <chr>                     <chr>           <POINT [°]> <dbl> <dbl>
#>  1 5     Zgłowiączka-Strózewo-Par… powyże… (18.65144 52.56443)  18.7  52.6
#>  2 8     DopZStarRadziejewa - Wit… Koloni… (18.67506 52.59173)  18.7  52.6
#>  3 4     Zgłąwiączka-ponizej. Osi… Samszy… (18.69878 52.60535)  18.7  52.6
#>  4 2     Zgłowiączka-pow.Osiecin.… powyże… (18.65691 52.64346)  18.7  52.6
#>  5 6     RowZ.K.Borucin-Samszyce   poniże… (18.71233 52.60872)  18.7  52.6
#>  6 7     rówZ.Osięcina             poniże…  (18.71868 52.6198)  18.7  52.6
#>  7 3     rów2Z.Osięcina            powyże… (18.70704 52.62894)  18.7  52.6
#>  8 13    oczyszcz kom. Rejów       wylot … (18.73169 52.62546)  18.7  52.6
#>  9 10    Zgłowiączka-Faliszewo     zastaw… (18.65371 52.57597)  18.7  52.6
#> 10 14    dren Karolin              duży w… (18.73097 52.62228)  18.7  52.6
#> 11 3'    Dren- pow.Osiecina        dren w… (18.70702 52.62893)  18.7  52.6
#> 12 11    rów z oczyszcz. Bytow     rów od…   (18.6503 52.5653)  18.7  52.6
#> 13 15    rów z Bolszewa            rów me… (18.68638 52.64021)  18.7  52.6
#> 14 16    rów melior w Pilichowie   rów me… (18.66879 52.65544)  18.7  52.7
#> 15 12    odpływ z drenów  Morzyce  duży o… (18.62654 52.57689)  18.6  52.6
#> 16 1     przepust za Skibinem      przepu… (18.58608 52.62548)  18.6  52.6
#> 17 2'    2' wylot z sieci drenars… wylot … (18.65922 52.64237)  18.7  52.6
#> 18 4'    4' wylot z sieci drenars… wylot … (18.69946 52.60647)  18.7  52.6
#> 19 8'    NA                        Koloni… (18.67431 52.59201)  18.7  52.6
#> 20 3''   Zgłowiaczka- pow.Osiecin  kanał … (18.70707 52.62892)  18.7  52.6
#> # … with abbreviated variable name ¹​Description
#> 
#> $data
#> # A tibble: 146,500 × 5
#>    Station DATE                Variables Values Source     
#>    <chr>   <dttm>              <chr>      <dbl> <chr>      
#>  1 1       2008-03-11 00:00:00 N-NH4       0.05 grab sample
#>  2 1       2008-03-25 00:00:00 N-NH4       0.05 grab sample
#>  3 1       2008-04-15 00:00:00 N-NH4       0.05 grab sample
#>  4 1       2008-04-28 00:00:00 N-NH4       0.05 grab sample
#>  5 1       2008-05-12 00:00:00 N-NH4       0.05 grab sample
#>  6 1       2008-05-23 00:00:00 N-NH4       0.05 grab sample
#>  7 1       2008-11-17 00:00:00 N-NH4       0.83 grab sample
#>  8 1       2008-12-01 00:00:00 N-NH4       0.25 grab sample
#>  9 1       2008-12-15 00:00:00 N-NH4       0.51 grab sample
#> 10 1       2009-03-09 00:00:00 N-NH4       1.89 grab sample
#> # … with 146,490 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>