Preparing interactive map of monitoring points with data inside

Preparing interactive map of monitoring points with data inside

Usage

plot_map(df, df_station, rch, shp)

Arguments

df

dataframe with formatted data (Station, DATE, Variables and Values columns are needed).

df_station

dataframe with formatted station data (ID, geometry columns are needed).

rch

sf dataframe for reaches (id, geometry columns are needed).

shp

sf dataframe for basin (name, geometry columns are needed).

Value

leaflet object of interactive map with monitoring data opening while pressing on points.

Examples

temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
#> [1] "Loading data from template."
#> [1] "Loading of data is finished."
plot_map(cal_data$data, cal_data$stations, rch, shp)