Aggregating values of single station to interactive monthly box plots

Aggregating values of single station to interactive monthly box plots

Usage

plot_monthly(df, station, drop_variables = c())

Arguments

df

dataframe with formatted data (Station, DATE, Variables and Values columns are needed)

station

character indicating station, which should be selected for figure.

drop_variables

optional parameter of character vector, which parameters should be excluded from figure.

Value

plotly object of interactive figure

Examples

temp_path <- system.file("extdata", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
#> [1] "Loading data from template."
#> [1] "Loading of data is finished."
plot_monthly(cal_data$data, station = "4", drop_variables = c("Q"))
#> Warning: The `keep` argument of `group_map()` is deprecated as of dplyr 1.0.0.
#> ℹ Please use the `.keep` argument instead.
#> ℹ The deprecated feature was likely used in the svatools package.
#>   Please report the issue at <https://github.com/biopsichas/svatools/issues>.