Aggregating values of single station to interactive monthly box plots

Aggregating values of single station to interactive monthly box plots

Usage

plot_monthly(df, station, variables = levels(as.factor(df$Variables)))

Arguments

df

dataframe with formatted data (Station, DATE, Variables and Values columns are needed)

station

character indicating station, which should be selected for figure.

variables

optional parameter of character vector, which parameters should be in figure.

Value

plotly object of interactive figure

Examples

temp_path <- system.file("extdata", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
#> [1] "Loading data from template."
#> [1] "Loading of data is finished."
plot_monthly(cal_data$data, station = "4")